"ads1_m14_a2: 2/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ assignments/m14/14.2 student reservation system/Student.java
@@ -117,7 +117,7 @@ class Student implements Comparable<Student> {
         String tokens[] = this.dob.split("-");
         return Integer.parseInt(tokens[2]);
     }
-    int getyear() {
+    long getyear() {
         String[] tokens = this.dob.split("-");
         return Integer.parseInt(tokens[3]);
     }
